pyproject.toml
[project]
name = "auto-labeler"
version = "0.1.0"
description = "A pragmatic AI-powered data labeling library."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.0.0",
    "tenacity",
    "scikit-learn",
    "numpy",
    "python-dotenv",
    "pyyaml",
    "click",
    "diskcache"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio"
]

[project.scripts]
auto-labeler = "auto_labeler.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_labeler"]
